Time: 0 ms (100%), Space: 19.8 MB (41.29%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -24,7 +24,7 @@ public:
             for( int i  = 0 ;i<size ; i++){
                 
                 auto node = q.front() ;
-                int initial = node.second-left ;
+                long long int initial = node.second-left ;
                 q.pop() ;
                 
                 if( node.first->left ) q.push( {node.first->left , initial*2+1 });
